requirements.txt
# Packages listed in alphabetical order

pyserial
